Revert "Refact/회원가입 페이지 컴포넌트로 쪼개기 (#237)"
This reverts commit 61db06c2773e63b8b2cffe4e2f9b7ee0b1cc516f.
