--- src/components/UserInput/styles.ts
@@ -1,9 +1,0 @@
-import { Text } from '@chakra-ui/react'
-import styled from '@emotion/styled'
-
-const RegisterText = styled(Text)`
-  margin-top: 1em;
-  margin-bottom: 0.5em;
-`
-
-export { RegisterText }
