+++ src/pages/Register/TermsOfUse/index.tsx
@@ -1,14 +1,24 @@
-import { memo, useCallback, useState } from 'react'
-import { useDispatch, useSelector } from 'react-redux'
-import TermsOfUseAccordion from '../../../components/Register/Accordion'
-import TermOfUseList from '../../../components/Register/TermOfUseList'
-import { setAgree } from '../../../redux/register/registerSlice'
-import { RootState } from '../../../redux/store'
-import { TermsOfUseBox } from './styles'
+import { CheckIcon } from '@chakra-ui/icons'
+import { Divider } from '@chakra-ui/react'
+import { useState } from 'react'
+import { setAgree } from '@redux/register/registerSlice'
+import {
+  TermsOfUseCollapse,
+  TermsOfUseCheckIcon,
+  TermsOfUseIcon,
+  List,
+  TermsOfUseBtn,
+  TermsOfUseCheckFlex,
+  TermsOfUseTotalBox,
+  TermsOfUseBox,
+} from './styles'
+import { useAppDispatch, useAppSelect } from '@redux/store.hooks'
 
-const TermsOfUse = memo((): JSX.Element => {
-  const agreeInfo = useSelector((state: RootState) => state.register.agreeInfo)
-  const dispatch = useDispatch()
+const TermsOfUse = (): JSX.Element => {
+  const [firstOpen, setFirstOpen] = useState(false)
+  const [secondOpen, setSecondOpen] = useState(false)
+  const agreeInfo = useAppSelect(state => state.register.agreeInfo)
+  const dispatch = useAppDispatch()
 
   const handleAgree = (value: string) => {
     switch (value) {
@@ -34,18 +44,87 @@ const TermsOfUse = memo((): JSX.Element => {
     }
   }
 
-  const handleClick = useCallback((e: any) => {
+  const handleClick = (e: any) => {
     const target = e.currentTarget
     const value = target.dataset.value
     value && handleAgree(value)
-  }, [])
-
+  }
   return (
     <TermsOfUseBox>
-      <TermOfUseList onClick={handleClick}></TermOfUseList>
-      <TermsOfUseAccordion onClick={handleClick}></TermsOfUseAccordion>
+      <List>
+        <TermsOfUseCheckFlex as="label">
+          <TermsOfUseCheckIcon
+            onClick={handleClick}
+            aria-label="Call Segun"
+            icon={<CheckIcon />}
+            data-value="total"
+            isclicked={
+              agreeInfo.private && agreeInfo.service ? 'true' : 'false'
+            }
+          />
+          <TermsOfUseTotalBox>전체 약관 동의</TermsOfUseTotalBox>
+        </TermsOfUseCheckFlex>
+      </List>
+      <Divider orientation="horizontal" borderBottomWidth="2.5px" />
+      <List>
+        <TermsOfUseCheckFlex as="label">
+          <TermsOfUseCheckIcon
+            aria-label="Call Segun"
+            icon={<CheckIcon />}
+            onClick={handleClick}
+            data-value="service"
+            isclicked={agreeInfo.service ? 'true' : 'false'}
+          />
+          <TermsOfUseBox>서비스 이용약관 동의</TermsOfUseBox>
+        </TermsOfUseCheckFlex>
+        <TermsOfUseBtn
+          onClick={() => {
+            if (secondOpen === true) {
+              setSecondOpen(!secondOpen)
+            }
+            setFirstOpen(!firstOpen)
+          }}
+        >
+          <TermsOfUseIcon></TermsOfUseIcon>
+        </TermsOfUseBtn>
+      </List>
+      <Divider orientation="horizontal" />
+      <TermsOfUseCollapse in={firstOpen}>
+        Lorem ipsum dolor sit, amet consectetur adipisicing elit. Dignissimos
+        at, quia officiis nesciunt quod blanditiis aperiam repellendus
+        praesentium numquam ullam maxime, vel consequuntur est dolore tempore
+        excepturi rerum ipsa perspiciatis?
+      </TermsOfUseCollapse>
+      <List>
+        <TermsOfUseCheckFlex as="label">
+          <TermsOfUseCheckIcon
+            aria-label="Call Segun"
+            icon={<CheckIcon />}
+            onClick={handleClick}
+            data-value="private"
+            isclicked={agreeInfo.private ? 'true' : 'false'}
+          />
+          <TermsOfUseBox>개인정보수집 및 이용동의</TermsOfUseBox>
+        </TermsOfUseCheckFlex>
+        <TermsOfUseBtn
+          onClick={() => {
+            if (firstOpen === true) {
+              setSecondOpen(!firstOpen)
+            }
+            setSecondOpen(!secondOpen)
+          }}
+        >
+          <TermsOfUseIcon></TermsOfUseIcon>
+        </TermsOfUseBtn>
+      </List>
+      <TermsOfUseCollapse in={secondOpen}>
+        Lorem ipsum dolor sit, amet consectetur adipisicing elit. Dignissimos
+        at, quia officiis nesciunt quod blanditiis aperiam repellendus
+        praesentium numquam ullam maxime, vel consequuntur est dolore tempore
+        excepturi rerum ipsa perspiciatis?
+      </TermsOfUseCollapse>
     </TermsOfUseBox>
   )
-})
+}
 
 export default TermsOfUse
