+++ src/pages/Register/index.tsx
@@ -1,32 +1,50 @@
-import React, { FormEvent, useCallback, useEffect, useState } from 'react'
+import React, { FormEvent, useEffect, useRef, useState } from 'react'
 import {
   postConfirmVerificationCode,
   postCreateVerificationCode,
 } from '@api/verificationAPI'
-import { Text, useToast } from '@chakra-ui/react'
+import { postDelete, postRegister } from '@api/userAPI'
+import {
+  chakra,
+  FormControl,
+  InputGroup,
+  InputLeftElement,
+  Stack,
+  Text,
+  Input,
+  Spinner,
+  Flex,
+  useToast,
+} from '@chakra-ui/react'
+import theme from '@style/theme'
+import { FaUserAlt, FaLock } from 'react-icons/fa'
 import {
   RegisterStack,
+  RegisterText,
   RegisterButton,
   ErrorText,
   DeleteBtn,
   Title,
 } from './styles'
 import { useNavigate } from 'react-router-dom'
+
 import TermsOfUse from './TermsOfUse'
 import { setWhichInfo } from '@redux/register/registerSlice'
-import LoadingSpinner from '@components/LoadingSpinner'
-import UserInput from '@components/UserInput'
 import { useAppDispatch, useAppSelect } from '@redux/store.hooks'
-import { postDelete, postRegister } from '@api/userAPI'
+
 const Register = (): JSX.Element => {
+  const CFaUserAlt = chakra(FaUserAlt)
+  const CFaLock = chakra(FaLock)
   const navigate = useNavigate()
+  const emailRef = useRef<HTMLInputElement>(null)
+  const verificationRef = useRef<HTMLInputElement>(null)
+  const passwordRef = useRef<HTMLInputElement>(null)
   const [input, setInput] = useState<RegisterForm>({
     username: '',
     password: '',
     verificationCode: '',
     passwordConfirm: '',
   })
-
   const toast = useToast({
     variant: 'toast',
   })
@@ -34,7 +52,8 @@ const Register = (): JSX.Element => {
   const whichUI = useAppSelect(state => state.register.whichUI)
   const agreeInfo = useAppSelect(state => state.register.agreeInfo)
   const [errorText, setErrorText] = useState('')
-  const handleChange = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     const { name, value } = e.target
     if (name === 'verificationCode') {
       if (value.length > 6) {
@@ -44,7 +63,7 @@ const Register = (): JSX.Element => {
       return
     }
     setInput({ ...input, [name]: value })
-  }, [])
+  }
 
   const handleAgreeBtn = () => {
     if (agreeInfo.private === true && agreeInfo.service === true) {
@@ -69,7 +88,7 @@ const Register = (): JSX.Element => {
     setErrorText(message)
     if (message === '이미 존재하는 회원입니다.') {
       setInput({ ...input, username: '' })
-      // emailRef.current && emailRef.current.focus()
+      emailRef.current && emailRef.current.focus()
       dispatch(setWhichInfo('Loading'))
     } else if (status < 400) {
       dispatch(setWhichInfo('Loading'))
@@ -93,7 +112,7 @@ const Register = (): JSX.Element => {
     switch (message) {
       case '존재하지 않는 인증 정보입니다.': {
         setInput({ ...input, verificationCode: '' })
-        // verificationRef.current && verificationRef.current.focus()
+        verificationRef.current && verificationRef.current.focus()
         break
       }
       case '만료된 인증 코드입니다.': {
@@ -152,7 +171,7 @@ const Register = (): JSX.Element => {
         dispatch(setWhichInfo('Valid'))
       }
     } else {
-      // passwordRef.current && passwordRef.current.focus()
+      passwordRef.current && passwordRef.current.focus()
       setErrorText('비밀번호가 일치하지 않습니다')
     }
   }
@@ -198,50 +217,86 @@ const Register = (): JSX.Element => {
           <Title>회원가입</Title>
           {!whichUI.isAgreed && <TermsOfUse></TermsOfUse>}
           {whichUI.isAgreed && (
-            <UserInput
-              text="이메일"
-              name="username"
-              type="email"
-              value={input.username}
-              placeholder="지스트 메일을 입력하세요"
-              onChange={handleChange}
-              disabled={whichUI.isCodeRequested}
-            ></UserInput>
+            <FormControl isRequired>
+              <RegisterText>이메일</RegisterText>
+              <InputGroup borderColor={`${theme.color.ligthGray}`}>
+                <InputLeftElement>
+                  {<CFaUserAlt color="gray.300" />}
+                </InputLeftElement>
+                <Input
+                  ref={emailRef}
+                  type="email"
+                  name="username"
+                  placeholder="지스트 메일을 입력하세요"
+                  value={input.username}
+                  onChange={handleChange}
+                  disabled={whichUI.isCodeRequested}
+                  borderRadius="0"
+                ></Input>
+              </InputGroup>
+            </FormControl>
           )}
 
           {whichUI.isCodeRequested && !whichUI.isExpired && (
-            <UserInput
-              text="인증 코드"
-              name="verificationCode"
-              type="text"
-              value={input.verificationCode}
-              placeholder="이메일로 온 인증 코드를 입력하세요"
-              onChange={handleChange}
-              disabled={whichUI.isVerificated}
-            ></UserInput>
+            <FormControl isRequired>
+              <Text mb="8px">인증 코드</Text>
+              <InputGroup borderColor={`${theme.color.ligthGray}`}>
+                <InputLeftElement>
+                  {<CFaLock color="gray.300" />}
+                </InputLeftElement>
+                <Input
+                  ref={verificationRef}
+                  type="text"
+                  name="verificationCode"
+                  placeholder="이메일로 온 인증 코드를 입력하세요"
+                  value={input.verificationCode}
+                  onChange={handleChange}
+                  disabled={whichUI.isVerificated}
+                  style={{ textTransform: 'uppercase' }}
+                  borderRadius="0"
+                ></Input>
+              </InputGroup>
+            </FormControl>
           )}
           {whichUI.isVerificated && (
-            <UserInput
-              text="비밀번호"
-              name="password"
-              type="password"
-              value={input.password}
-              placeholder="영문과 숫자를 포함한 8자리 이상의 비밀번호를 입력하세요"
-              onChange={handleChange}
-            ></UserInput>
+            <FormControl isRequired>
+              <Text mb="8px">비밀번호</Text>
+              <InputGroup borderColor={`${theme.color.ligthGray}`}>
+                <InputLeftElement>
+                  {<CFaLock color="gray.300" />}
+                </InputLeftElement>
+                <Input
+                  ref={passwordRef}
+                  type="password"
+                  name="password"
+                  placeholder="영문과 숫자를 포함한 8자리 이상의 비밀번호를 입력하세요"
+                  value={input.password}
+                  onChange={handleChange}
+                  borderRadius="0"
+                ></Input>
+              </InputGroup>
+            </FormControl>
           )}
           {whichUI.isVerificated && (
-            <UserInput
-              text="비밀번호 확인"
-              name="passwordConfirm"
-              type="password"
-              value={input.passwordConfirm}
-              placeholder="비밀번호를 재입력하세요"
-              onChange={handleChange}
-            ></UserInput>
+            <FormControl isRequired>
+              <Text mb="8px">비밀번호 확인</Text>
+              <InputGroup borderColor={`${theme.color.ligthGray}`}>
+                <InputLeftElement>
+                  {<CFaLock color="gray.300" />}
+                </InputLeftElement>
+                <Input
+                  type="password"
+                  name="passwordConfirm"
+                  placeholder="비밀번호를 재입력하세요"
+                  value={input.passwordConfirm}
+                  onChange={handleChange}
+                  borderRadius="0"
+                ></Input>
+              </InputGroup>
+            </FormControl>
           )}
           {!whichUI.isAgreed && (
-            <RegisterButton onClick={handleAgreeBtn}>다음 단계</RegisterButton>
+            <RegisterButton onClick={handleAgreeBtn}>다음단계</RegisterButton>
           )}
           {whichUI.isAgreed &&
             !whichUI.isCodeRequested &&
@@ -256,7 +311,22 @@ const Register = (): JSX.Element => {
               인증코드 재전송
             </RegisterButton>
           )}
-          {whichUI.isLoading && <LoadingSpinner></LoadingSpinner>}
+
+          {whichUI.isLoading && (
+            <Flex flexDirection="column" alignItems="center">
+              <Spinner
+                m="auto"
+                thickness="4px"
+                speed="0.65s"
+                emptyColor="gray.200"
+                color="#df3127"
+                size="xl"
+                mb="10px"
+              />
+              잠시만 기다려주세요...
+            </Flex>
+          )}
+
           {whichUI.isCodeRequested &&
             !whichUI.isVerificated &&
             !whichUI.isExpired && (
