+++ src/pages/Register/styles.ts
@@ -34,6 +34,11 @@ const RegisterButton = styled.button`
   font-weight: bold;
 `
 
+const RegisterText = styled(Text)`
+  margin-top: 1em;
+  margin-bottom: 0.5em;
+`
+
 const ErrorText = styled.p`
   margin-top: 2em;
   text-align: center;
@@ -51,4 +56,11 @@ const DeleteBtn = styled.button`
   font-weight: bold;
 `
 
-export { RegisterStack, Title, RegisterButton, ErrorText, DeleteBtn }
+export {
+  RegisterStack,
+  Title,
+  RegisterText,
+  RegisterButton,
+  ErrorText,
+  DeleteBtn,
+}
