+++ src/types/user.d.ts
@@ -46,11 +46,6 @@ interface RegisterAgree {
 interface AgreeAttribute {
   isClicked: boolean
 }
-
-interface TermOfUseList {
-  onClick: any
-}
-
 // 회원 가입 UI 관련
 
 interface FindPasswordWhichUI {
